+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- SearchMetadata for category, brand and collection.
+
 ## [1.28.4] - 2020-12-07
 
 ### Fixed